{"schema":"decodex/commit/1","summary":"Split consolidation proposal storage modules","authority":"manual"}
